# brick_judge/openapi.json
{"openapi":"3.1.0","info":{"title":"Brick Judge API","description":"\n## 🧱 LDR 브릭 구조 물리 검증 API\n\nLEGO/브릭 모델의 구조적 안정성을 검증하는 API입니다.\n\n### 주요 기능\n- **물리 검증**: 브릭 구조의 안정성 분석\n- **이슈 탐지**: floating(공중 부유), isolated(고립), top_only(위에서만 연결) 등\n- **점수 산출**: 0-100점 안정성 점수\n\n### LDR 포맷\nLDraw 표준 포맷 (.ldr, .dat)을 사용합니다.\n```\n0 Model Name\n1 <color> <x> <y> <z> <rotation matrix 9개> <part>.dat\n```\n\n### 사용 예시 (LLM용)\n1. `POST /api/judge` 에 LDR 문자열 전송\n2. 점수와 이슈 목록 확인\n3. 점수 50 미만이면 구조 수정 필요\n","contact":{"name":"Brickers Team","url":"https://github.com/sio60/brickers-ai"},"license":{"name":"MIT"},"version":"1.0.0"},"paths":{"/":{"get":{"summary":"Index","operationId":"index__get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/api/status":{"get":{"tags":["info"],"summary":"서버 상태 확인","description":"서버 상태와 백엔드 정보를 반환합니다.","operationId":"status_api_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/verify":{"post":{"tags":["viewer"],"summary":"LDR 파일 업로드 검증","description":"파일 업로드 방식의 LDR 검증 (웹 UI용)","operationId":"verify_ldr_api_verify_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_verify_ldr_api_verify_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/info":{"get":{"tags":["info"],"summary":"백엔드 정보","description":"사용 중인 물리 엔진 백엔드 정보를 반환합니다.","operationId":"info_api_info_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BackendInfo"}}}}}}},"/viewer":{"get":{"summary":"Viewer","description":"3D 뷰어 페이지","operationId":"viewer_viewer_get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/api/judge":{"post":{"tags":["judge"],"summary":"🎯 LDR 브릭 구조 물리 검증 (LLM용 메인 API)","description":"## LDR 브릭 구조의 물리적 안정성을 검증합니다.\n\n### 입력\n- `ldr_content`: LDraw 포맷 문자열\n\n### 출력\n- `score`: 0-100점 (50점 이상이면 안정)\n- `stable`: 안정 여부\n- `issues`: 발견된 문제 목록\n\n### 이슈 타입\n- `floating`: 공중에 떠있는 브릭 (critical)\n- `isolated`: 다른 브릭과 연결되지 않음 (high)\n- `top_only`: 위에서만 연결됨, 아래 지지 없음 (medium)\n\n### 사용 예시\n```python\nresponse = requests.post(\"/api/judge\", json={\n    \"ldr_content\": \"0 My Model\\n1 4 0 0 0 1 0 0 0 1 0 0 0 1 3001.dat\"\n})\nif response.json()[\"score\"] < 50:\n    print(\"구조 수정 필요!\")\n```","operationId":"judge_ldr_api_judge_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LdrRequest"}}},"required":true},"responses":{"200":{"description":"검증 결과: 점수, 안정성, 이슈 목록","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JudgeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"BackendInfo":{"properties":{"backend":{"type":"string","title":"Backend","description":"백엔드 타입","example":"rust"},"version":{"type":"string","title":"Version","description":"버전","example":"0.1.0"},"module":{"type":"string","title":"Module","description":"모듈명","example":"brick_judge_rs"}},"type":"object","required":["backend","version","module"],"title":"BackendInfo","description":"백엔드 정보"},"Body_verify_ldr_api_verify_post":{"properties":{"file":{"type":"string","format":"binary","title":"File","description":"LDR 파일 (.ldr, .dat)"}},"type":"object","required":["file"],"title":"Body_verify_ldr_api_verify_post"},"BrickIssue":{"properties":{"brick_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Brick Id","description":"문제가 있는 브릭 ID (0부터 시작)"},"type":{"type":"string","title":"Type","description":"이슈 타입","example":"floating"},"severity":{"type":"string","title":"Severity","description":"심각도: critical, high, medium, low","example":"critical"},"message":{"type":"string","title":"Message","description":"이슈 설명","example":"브릭 #5 바닥과 연결 안됨"},"color":{"type":"string","title":"Color","description":"시각화용 색상 (hex)","example":"#FF0000"}},"type":"object","required":["type","severity","message","color"],"title":"BrickIssue","description":"브릭 구조 이슈"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"JudgeResponse":{"properties":{"model_name":{"type":"string","title":"Model Name","description":"모델 이름","example":"Simple Tower"},"brick_count":{"type":"integer","title":"Brick Count","description":"총 브릭 개수","example":15},"score":{"type":"integer","title":"Score","description":"안정성 점수 (0-100, 50 이상이면 안정)","example":85},"stable":{"type":"boolean","title":"Stable","description":"안정 여부 (score >= 50 and no critical)","example":true},"issues":{"items":{"$ref":"#/components/schemas/BrickIssue"},"type":"array","title":"Issues","description":"발견된 이슈 목록","default":[]},"brick_colors":{"additionalProperties":{"type":"string"},"type":"object","title":"Brick Colors","description":"이슈 브릭별 색상 맵","default":{}},"elapsed_ms":{"type":"number","title":"Elapsed Ms","description":"처리 시간 (밀리초)","example":1.23},"backend":{"type":"string","title":"Backend","description":"사용된 백엔드 엔진","example":"rust"}},"type":"object","required":["model_name","brick_count","score","stable","elapsed_ms","backend"],"title":"JudgeResponse","description":"물리 검증 결과","example":{"backend":"rust","brick_colors":{"5":"#0055FF"},"brick_count":15,"elapsed_ms":1.23,"issues":[{"brick_id":5,"color":"#0055FF","message":"브릭 #5 위에서만 연결됨","severity":"medium","type":"top_only"}],"model_name":"Simple Tower","score":70,"stable":true}},"LdrRequest":{"properties":{"ldr_content":{"type":"string","title":"Ldr Content","description":"LDraw 포맷의 브릭 모델 문자열","example":"0 Simple Tower\n1 4 0 0 0 1 0 0 0 1 0 0 0 1 3001.dat\n1 4 0 -24 0 1 0 0 0 1 0 0 0 1 3001.dat\n1 1 0 -48 0 1 0 0 0 1 0 0 0 1 3003.dat"}},"type":"object","required":["ldr_content"],"title":"LdrRequest","description":"LDR 검증 요청"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}},"tags":[{"name":"judge","description":"브릭 구조 물리 검증 (LLM용 메인 API)"},{"name":"info","description":"서버 상태 및 정보"},{"name":"viewer","description":"웹 UI"}]}